Move around oauth file structure
